Split tests into unit and integration suites
